Moved summary efile view for reports to the same file as other reports
